Ported XtraBackup to MySQL 5.1.55, MySQL 5.5.9 and Percona Server 5.1.55-rel12.6 codebases.
This revision also fixes the following build- and test-related bugs:

bug #733811
bug #733820
bug #733830
bug #733842

+++ test/t/xb_export.sh
@@ -2,7 +2,11 @@
 
 init
 
-backup_dir=/tmp/xb_export_backup
+if [ "$MYSQL_VERSION" != "percona" ]; then
+  exit $SKIPPED_EXIT_CODE
+fi
+
+backup_dir=$topdir/xb_export_backup
 rm -rf $backup_dir
 mkdir $backup_dir
 # Starting database server
@@ -28,21 +32,19 @@ vlog "rowsnum_1 is $rowsnum_1"
 vlog "checksum_1 is $checksum_1"
 
 # Performing table backup
-#run_cmd xtrabackup --datadir=$mysql_datadir --backup --tables="^incremental_sample[.]test" --target-dir=$backup_dir
-run_cmd xtrabackup --datadir=$mysql_datadir --backup --target-dir=$backup_dir
+run_cmd xtrabackup --no-defaults --datadir=$mysql_datadir --backup --target-dir=$backup_dir
 vlog "Table was backed up"
 
 vlog "Re-initializing database server"
 stop_mysqld
-clean
-init
+# Can't use clean/init because that will remove $backup_dir as well
+clean_datadir
 run_mysqld --innodb_file_per_table --innodb_expand_import=1 --innodb_file_format=Barracuda
 load_dbase_schema incremental_sample
 vlog "Database was re-initialized"
 
 run_cmd ${MYSQL} ${MYSQL_ARGS} -e "alter table test discard tablespace;" incremental_sample
-#run_cmd xtrabackup --datadir=$mysql_datadir --prepare --target-dir=$backup_dir
-run_cmd xtrabackup --datadir=$mysql_datadir --prepare --export --target-dir=$backup_dir
+run_cmd xtrabackup --no-defaults --datadir=$mysql_datadir --prepare --export --target-dir=$backup_dir
 run_cmd cp $backup_dir/incremental_sample/test* $mysql_datadir/incremental_sample/
 run_cmd ls -lah $mysql_datadir/incremental_sample/
 run_cmd ${MYSQL} ${MYSQL_ARGS} -e "alter table test import tablespace" incremental_sample
@@ -69,9 +71,9 @@ run_cmd ${MYSQL} ${MYSQL_ARGS} -e "show create table test;" incremental_sample
 # Performing full backup of imported table
 mkdir -p $topdir/backup/full
 
-xtrabackup --datadir=$mysql_datadir --backup --target-dir=$topdir/backup/full
-xtrabackup --datadir=$mysql_datadir --prepare --apply-log-only --target-dir=$topdir/backup/full
-xtrabackup --datadir=$mysql_datadir --prepare --target-dir=$topdir/backup/full
+xtrabackup --no-defaults --datadir=$mysql_datadir --backup --target-dir=$topdir/backup/full
+xtrabackup --no-defaults --datadir=$mysql_datadir --prepare --apply-log-only --target-dir=$topdir/backup/full
+xtrabackup --no-defaults --datadir=$mysql_datadir --prepare --target-dir=$topdir/backup/full
 
 run_cmd ${MYSQL} ${MYSQL_ARGS} -e "delete from test;" incremental_sample
